Change CryoPodSystem: Move code from server to shared
This is still without adding UI prediction
